docs: Make PIL example more in line with the others
The example adds an alpha channel to the image if it doesn't have one,
which is more of an example about PIL than cairo-PIL integration.
(Aside from that, it modified the argument image, which is a bad
practice, and just ignored the alpha parameter if the image already had
an alpha channel)

Instead, just convert the image to a surface of the matching format.